Refactor project and add new material
