+++ docker/run.sh
@@ -1,0 +1,5 @@
+cd ..
+
+docker build -t jupyter-app -f docker/Dockerfile .
+
+docker run -p 8080:8080 jupyter-app
